+++ CHANGELOG.rst
@@ -11,3 +11,4 @@ malformed XML instead of an actual erroneous condition.
 * Added: Support of Container Security Containers API.
 * Added: Support of Container Security Test Jobs API.
 * Added: Support of Container Security Reports API.
+* Added: Support of Container Security Policy API.